Heavily modified code to have container classes which holds the data structures, wrote the tests for these as well
