+++ lib/puppy_breeder/purchase_request.rb
@@ -3,13 +3,10 @@ module PuppyBreeder
   class PurchaseRequest
     attr_reader :breed, :id
     attr_accessor :status
-    @@counter = 0
 
     def initialize(breed)
       @breed = breed
       @status = 'pending'
-      @@counter += 1
-      @id = @@counter
     end
 
     def complete_order
